+++ src/ui_basic/listselect.cc
@@ -362,9 +362,9 @@ void BaseListselect::draw(RenderTarget& dst) {
 	if (selection_mode_ == ListselectLayout::kDropdown) {
 		RGBAColor black(0, 0, 0, 255);
 		//  top edge
-		dst.brighten_rect(Rectf(0.f, 0.f, get_w(), 2.f), BUTTON_EDGE_BRIGHT_FACTOR / 4);
+		dst.brighten_rect(Recti(0, 0, get_w(), 2), BUTTON_EDGE_BRIGHT_FACTOR / 4);
 		//  left edge
-		dst.brighten_rect(Rectf(0.f, 0.f, 2.f, get_h()), BUTTON_EDGE_BRIGHT_FACTOR);
+		dst.brighten_rect(Recti(0, 0, 2, get_h()), BUTTON_EDGE_BRIGHT_FACTOR);
 		//  bottom edge
 		dst.fill_rect(Rectf(2.f, get_h() - 2.f, get_eff_w() - 2.f, 1.f), black);
 		dst.fill_rect(Rectf(1.f, get_h() - 1.f, get_eff_w() - 1.f, 1.f), black);
@@ -372,7 +372,7 @@ void BaseListselect::draw(RenderTarget& dst) {
 		dst.fill_rect(Rectf(get_w() - 2.f, 1.f, 1.f, get_h() - 1.f), black);
 		dst.fill_rect(Rectf(get_w() - 1.f, 0.f, 1.f, get_h()), black);
 	} else {
-		dst.brighten_rect(Rectf(0.f, 0.f, get_eff_w(), get_h()), ms_darken_value);
+		dst.brighten_rect(Recti(0, 0, get_eff_w(), get_h()), ms_darken_value);
 	}
 
 	while (idx < entry_records_.size()) {
@@ -397,7 +397,7 @@ void BaseListselect::draw(RenderTarget& dst) {
 
 		// Highlight the current selected entry
 		if (idx == selection_) {
-			Rectf r(point, maxw, lineheight_);
+			Recti r(point, maxw, lineheight_);
 			if (r.x < 0) {
 				r.w += r.x;
 				r.x = 0;
